requirements.txt
bootstrap-admin==0.4.4
certifi==2021.10.8
charset-normalizer==2.0.8
defusedxml==0.7.1
Django==2.2
django-cors-headers==3.8.0
django-environ==0.8.1
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
lxml==4.6.4
Markdown==3.3.4
psycopg2==2.8.6
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
celery